ForSyDe-SystemC: fixed a warning related to signed vs unsigned comparison

+++ ForSyDe-SystemC/trunk/src/forsyde/sy_process_constructors_strict.hpp
@@ -466,7 +466,7 @@ private:
 
     void prep()
     {
-        for (unsigned int i=0; i<N; i++)
+        for (size_t i=0; i<N; i++)
         {
             auto ival_temp = iport.read();
             CHECK_PRESENCE(ival_temp);
@@ -493,7 +493,7 @@ private:
     void bindInfo()
     {
         boundInChans.resize(N);     // only one input port
-        for (unsigned int i=0; i<N; i++)
+        for (size_t i=0; i<N; i++)
         	boundInChans[i].port = &iport[i];
         boundOutChans.resize(1);    // only one output port
         boundOutChans[0].port = &oport1;
@@ -635,7 +635,7 @@ private:
 
     void exec()
     {
-        for (unsigned int i=0; i<N; i++)
+        for (size_t i=0; i<N; i++)
         {
             _func(oval[i], ival[i]);
         }
@@ -1447,7 +1447,7 @@ private:
     
     void prep()
     {
-        for (int i=0; i<N; i++)
+        for (size_t i=0; i<N; i++)
         {
             auto ival_temp = iport[i].read();
             CHECK_PRESENCE(ival_temp);
@@ -1469,7 +1469,7 @@ private:
     void bindInfo()
     {
         boundInChans.resize(N);     // N input ports
-        for (int i=0;i<N;i++)
+        for (size_t i=0;i<N;i++)
             boundInChans[i].port = &iport[i];
         boundOutChans.resize(1);    // only one output port
         boundOutChans[0].port = &oport1;
@@ -1660,7 +1660,7 @@ private:
     
     void prod()
     {
-        for (int i=0; i<N; i++)
+        for (size_t i=0; i<N; i++)
             WRITE_MULTIPORT(oport[i],abst_ext<T1>(in_val->unsafe_from_abst_ext()[i]))  // write to the output i
     }
     
@@ -1675,7 +1675,7 @@ private:
         boundInChans.resize(1);     // only one input port
         boundInChans[0].port = &iport1;
         boundOutChans.resize(N);    // two output ports
-        for (int i=0;i<N;i++)
+        for (size_t i=0;i<N;i++)
             boundOutChans[i].port = &oport[i];
     }
 #endif
